:fire: move plug handling to its own file
